+++ src/dreamstate/core/storing/ActionsStore.ts
@@ -2,5 +2,12 @@
  * Class for nested action values.
  * Used by context diff checker on updates by ContextManager class.
  */
-export class ActionsStore {
+import { TAnyObject } from "@/dreamstate/types";
+
+export class ActionsStore<T extends TAnyObject = TAnyObject> {
+
+  public constructor(actions: T) {
+    Object.assign(this, actions);
+  }
+
 }