+++ src/dreamstate/test-utils.ts
@@ -1,12 +1,12 @@
 /*
  * 'https://github.com/Neloreck/dreamstate'
  */
-export * from "@/dreamstate/test-utils/registry/getCurrent";
-export * from "@/dreamstate/test-utils/utils/getReactConsumer";
-export * from "@/dreamstate/test-utils/utils/getReactProvider";
 
+export * from "@/dreamstate/test-utils/registry/getCurrent";
 export * from "@/dreamstate/test-utils/registry/mockManagerWithScope";
 export * from "@/dreamstate/test-utils/registry/mockScopeProvider";
 export * from "@/dreamstate/test-utils/registry/mockScope";
 
+export * from "@/dreamstate/test-utils/utils/getReactConsumer";
+export * from "@/dreamstate/test-utils/utils/getReactProvider";
 export * from "@/dreamstate/test-utils/utils/nextAsyncQueue";