+++ src/dreamstate/utils/createNested.test.ts
@@ -6,15 +6,17 @@ describe("Nested util", () => {
   it("Should properly create nested objects", () => {
     const nested: TNested<{ test: boolean }> = createNested({ test: true });
 
-    expect(Object.keys(nested)).toHaveLength(2);
+    expect(Object.keys(nested)).toHaveLength(1);
     expect(nested.test).toBeTruthy();
+    expect(nested.asMerged).toBeDefined();
   });
 
   it("Should properly mutate nested objects", () => {
     const nested: TNested<{ test: boolean }> = createNested({ test: true });
     const next: TNested<{ test: boolean }> = nested.asMerged({ test: false });
 
-    expect(Object.keys(next)).toHaveLength(2);
+    expect(Object.keys(next)).toHaveLength(1);
+    expect(next.asMerged).toBeDefined();
     expect(next.test).toBeFalsy();
   });