+++ src/dreamstate/utils/createNested.ts
@@ -2,16 +2,6 @@ import { NestedStore } from "@/dreamstate/core/storing/NestedStore";
 import { TAnyObject, TNested } from "@/dreamstate/types";
 
 /**
- * Util for nested stores immutable merging.
- */
-function asMerged<T extends TAnyObject>(
-  this: TNested<T>,
-  state: Partial<T> = {}
-): TNested<T> {
-  return Object.assign(new NestedStore(), this, state);
-}
-
-/**
  * Create nested sub-state for deeper shallow checking.
  * Useful when your context has some nested objects that should be checked separately.
  *
@@ -26,5 +16,5 @@ export function createNested<T extends TAnyObject>(
     throw new TypeError("Nested values can be created for non-null object parameter only.");
   }
 
-  return Object.assign(new NestedStore(), initialValue, { asMerged: asMerged as any });
+  return Object.assign(new NestedStore<T>(), initialValue);
 }